+++ tests/testthat/test-new-model.R
@@ -1,11 +1,11 @@
 context("Testing function to create or read in model object")
 
 
-test_that("read_model() returns expected object", {
+test_that("read_model() returns expected object [BBR-NWMD-001]", {
   expect_equal(read_model(MOD1_PATH), REF_LIST_1)
 })
 
-test_that("read_model() returns expected object from no ext specified", {
+test_that("read_model() returns expected object from no ext specified [BBR-NWMD-002]", {
   temp_path <- file.path(ABS_MODEL_DIR, "temp.yaml")
   ctl_path <- fs::path_ext_set(temp_path, "ctl")
 
@@ -21,7 +21,7 @@ test_that("read_model() returns expected object from no ext specified", {
   expect_equal(mod2, REF_LIST_TMP)
 })
 
-test_that("read_model() can read a model whose path has a period", {
+test_that("read_model() can read a model whose path has a period [BBR-NWMD-003]", {
   temp_ctl <- tempfile(pattern = "file.", fileext = ".ctl")
   # ensure that `temp_ctl` exists
   readr::write_file("foo", temp_ctl)
@@ -34,7 +34,7 @@ test_that("read_model() can read a model whose path has a period", {
   expect_identical(class(mod), NM_MOD_CLASS_LIST)
 })
 
-test_that("new_model() creates new YAML file", {
+test_that("new_model() creates new YAML file [BBR-NWMD-004]", {
   temp_mod_path <- create_temp_model()
   fs::file_delete(fs::path_ext_set(temp_mod_path, "yaml"))
 
@@ -51,11 +51,11 @@ test_that("new_model() creates new YAML file", {
   expect_true(all(MODEL_REQ_KEYS %in% names(mod1b)))
 })
 
-test_that("new_model() throws an error if the model file does not exist", {
+test_that("new_model() throws an error if the model file does not exist [BBR-NWMD-005]", {
   expect_error(new_model("foo", "bar"), "No model file found")
 })
 
-test_that("compare read_model() and new_model() objects", {
+test_that("compare read_model() and new_model() objects [BBR-NWMD-006]", {
   temp_mod_path <- create_temp_model()
   fs::file_delete(fs::path_ext_set(temp_mod_path, "yaml"))
 
@@ -87,7 +87,7 @@ test_that("compare read_model() and new_model() objects", {
   }
 })
 
-test_that("new_model() .overwrite arg works", {
+test_that("new_model() .overwrite arg works [BBR-NWMD-007]", {
   temp_mod_path <- create_temp_model()
 
   # error if file exists
@@ -104,7 +104,7 @@ test_that("new_model() .overwrite arg works", {
 })
 
 
-test_that("new_model() .based_on arg works", {
+test_that("new_model() .based_on arg works [BBR-NWMD-008]", {
   temp_mod_path <- create_temp_model()
   parent_model_id <- get_model_id(create_temp_model())
   fs::file_delete(fs::path_ext_set(temp_mod_path, "yaml"))
@@ -118,7 +118,7 @@ test_that("new_model() .based_on arg works", {
   expect_equal(mod1a[[YAML_BASED_ON]], parent_model_id)
 })
 
-test_that("new_model() .based_on arg errors on fake model", {
+test_that("new_model() .based_on arg errors on fake model [BBR-NWMD-009]", {
   temp_mod_path <- ctl_ext(tempfile())
   writeLines("CREATED BY: new_model() .based_on arg errors on fake model", temp_mod_path)
   on.exit(fs::file_delete(temp_mod_path))
@@ -134,7 +134,7 @@ test_that("new_model() .based_on arg errors on fake model", {
   )
 })
 
-test_that("new_model() supports `.path` containing a period", {
+test_that("new_model() supports `.path` containing a period [BBR-NWMD-010]", {
   temp_ctl <- tempfile(pattern = "file.", fileext = ".ctl")
   # ensure that `temp_ctl` exists
   readr::write_file("foo", temp_ctl)