+++ tests/testthat/test-param-estimates.R
@@ -11,12 +11,12 @@ ref_df1 <- dget(PARAM_REF_FILE)
 
 withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
 
-  test_that("param_estimates.bbi_model_summary gets expected table", {
+  test_that("param_estimates.bbi_model_summary gets expected table [BBR-PEST-001]", {
     par_df <- MOD1 %>% model_summary() %>% param_estimates()
     expect_equal(par_df, ref_df1)
   })
 
-  test_that("param_estimates correctly errors on Bayesian model", {
+  test_that("param_estimates correctly errors on Bayesian model [BBR-PEST-002]", {
     mod1 <- read_model(file.path(MODEL_DIR_X, "1001"))
     sum1 <- model_summary(mod1, .bbi_args = list(ext_file = "1001.1.TXT"))
     expect_error(
@@ -25,7 +25,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     )
   })
 
-  test_that("param_estimates correctly errors on Bayesian model with multiple estimation methods", {
+  test_that("param_estimates correctly errors on Bayesian model with multiple estimation methods [BBR-PEST-003]", {
     sum1 <- file.path(MODEL_DIR_X, "acop-fake-bayes") %>%
       read_model() %>%
       model_summary()
@@ -35,7 +35,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     )
   })
 
-  test_that("param_estimates correctly warns on mixture model", {
+  test_that("param_estimates correctly warns on mixture model [BBR-PEST-004]", {
     mod1 <- read_model(file.path(MODEL_DIR_X, "iovmm"))
     sum1 <- model_summary(mod1)