+++ tests/testthat/test-print.R
@@ -10,29 +10,29 @@ model_1 <- MOD1
 withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
   withr::local_envvar(c("NO_COLOR" = "true"))
 
-  test_that("print.bbi_process works with .wait = TRUE", {
+  test_that("print.bbi_process works with .wait = TRUE [BBR-PRNT-001]", {
     proc <- bbi_exec("--help", .wait = TRUE)
     res <- capture.output(print(proc))
-    expect_true(any(str_detect(res, PROC_HELP_STR)))
+    expect_true(any(str_detect(res, fixed(PROC_HELP_STR))))
     expect_true(any(str_detect(res, "Process finished.")))
   })
 
-  test_that("print.bbi_process works with .wait = FALSE", {
+  test_that("print.bbi_process works with .wait = FALSE [BBR-PRNT-001]", {
     proc <- bbi_exec("--help", .wait = FALSE)
     res <- capture.output(print(proc))
 
-    expect_true(any(str_detect(res, PROC_HELP_STR)))
+    expect_true(any(str_detect(res, fixed(PROC_HELP_STR))))
     expect_true(any(str_detect(res, "Not waiting for process to finish.")))
   })
 
-  test_that("print.bbi_process works with dry run", {
+  test_that("print.bbi_process works with dry run [BBR-PRNT-001]", {
     proc <- bbi_dry_run("--help", ".")
     res <- capture.output(print(proc))
-    expect_true(any(str_detect(res, PROC_HELP_STR)))
+    expect_true(any(str_detect(res, fixed(PROC_HELP_STR))))
     expect_true(any(str_detect(res, "DRY RUN! Process not actually run.")))
   })
 
-  test_that("print.bbi_process(.call_limit) works", {
+  test_that("print.bbi_process(.call_limit) works [BBR-PRNT-001]", {
     # create a bunch of fake models
     temp_dir <- file.path(get_model_working_directory(MOD1), "print-test")
     fs::dir_create(temp_dir)
@@ -40,20 +40,36 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     mods <- purrr::map(1:50, ~copy_model_from(MOD1, file.path("print-test", .x)))
     proc <- submit_models(mods, .dry_run = TRUE)
 
-    # check that default has bbi path, at least two model paths, and flags
-    call_str <- capture.output(print(proc[[1]]))[2]
-    expect_true(str_detect(call_str, read_bbi_path()))
-    expect_true(str_detect(call_str, as.character(glue("{temp_dir}/1\\.ctl.+{temp_dir}/2\\.ctl"))))
+    # Check that default has bbi path, at least two model paths, and flags.
+    # While the intention is to check the _default_ behavior, increase
+    # .call_site if needed to avoid this test failing on a system where the
+    # model paths are longer than expected.
+    args <- list(proc[[1]])
+    len_mod <- sum(nchar(purrr::map_chr(mods[1:2], get_model_path)),
+                   # +1 for space in between
+                   1)
+    if (len_mod > 250) {
+      args <- c(args, .call_limit = len_mod)
+    }
+
+    call_str <- capture.output(do.call(print, args))[2]
+    expect_true(str_detect(call_str, fixed(read_bbi_path())))
+    expect_true(str_detect(call_str,
+                           paste0("\\Q", temp_dir, "\\E", "/1\\.ctl.+",
+                                  "\\Q", temp_dir, "\\E", "/2\\.ctl.+")))
     expect_true(str_detect(call_str, "--overwrite --threads=4"))
 
-    # check that passing in .call_limit=30 has bbi path, NO model paths, but still has flags
+    # Check that passing in .call_limit=30 has bbi path, NO model paths, but
+    # still has flags. Note that 30 is sufficient to trigger truncation of the
+    # model arguments because "inst/model/nonmem/basic/print-test" alone is over
+    # 30 characters.
     call_str <- capture.output(print(proc[[1]], .call_limit=30))[2]
-    expect_true(str_detect(call_str, read_bbi_path()))
-    expect_false(str_detect(call_str, temp_dir))
+    expect_true(str_detect(call_str, fixed(read_bbi_path())))
+    expect_false(str_detect(call_str, fixed(temp_dir)))
     expect_true(str_detect(call_str, "--overwrite --threads=4"))
   })
 
-  test_that("print.bbi_nonmem_model contains proper fields if all present", {
+  test_that("print.bbi_nonmem_model contains proper fields if all present [BBR-PRNT-002]", {
     model_1[[YAML_NOTES]] <- c("x", "y")
 
     fields <- c('Status',
@@ -68,7 +84,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     })
   })
 
-  test_that("print.bbi_nonmem_model run status functions properly", {
+  test_that("print.bbi_nonmem_model run status functions properly [BBR-PRNT-002]", {
     bullets <- capture.output({ # these get thrown away, but we don't want them to print in the test output
       expect_message(print(model_1), regexp = "Finished Running")
 
@@ -81,7 +97,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     })
   })
 
-  test_that("print.bbi_nonmem_summary works basic FOCE model", {
+  test_that("print.bbi_nonmem_summary works basic FOCE model [BBR-PRNT-003]", {
     .s <- file.path(MODEL_DIR, 1) %>%
       read_model() %>%
       model_summary()
@@ -90,7 +106,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     expect_equal(res_str, ref_str)
   })
 
-  test_that("print.bbi_nonmem_summary works mixture model", {
+  test_that("print.bbi_nonmem_summary works mixture model [BBR-PRNT-003]", {
     .s <- file.path(MODEL_DIR_X, "iovmm") %>%
         read_model() %>%
         model_summary()
@@ -101,7 +117,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     expect_equal(res_str, ref_str)
   })
 
-  test_that("print.bbi_nonmem_summary works Bayes model", {
+  test_that("print.bbi_nonmem_summary works Bayes model [BBR-PRNT-003]", {
     .s <- file.path(MODEL_DIR_X, "1001") %>%
       read_model() %>%
       model_summary(.bbi_args = list(ext_file = "1001.1.TXT"))
@@ -110,7 +126,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     expect_equal(res_str, ref_str)
   })
 
-  test_that("print.bbi_nonmem_summary works SAEM-IMP model", {
+  test_that("print.bbi_nonmem_summary works SAEM-IMP model [BBR-PRNT-003]", {
     .s <- file.path(MODEL_DIR_X, "example2_saemimp") %>%
       read_model() %>%
       model_summary()
@@ -119,7 +135,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     expect_equal(res_str, ref_str)
   })
 
-  test_that("print.bbi_nonmem_summary works IOV model", {
+  test_that("print.bbi_nonmem_summary works IOV model [BBR-PRNT-003]", {
     # load a model summary
     .s <- file.path(MODEL_DIR_X, "acop-iov") %>%
       read_model() %>%
@@ -129,7 +145,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     expect_equal(res_str, ref_str)
   })
 
-  test_that("print.bbi_nonmem_summary .fixed=TRUE", {
+  test_that("print.bbi_nonmem_summary .fixed=TRUE [BBR-PRNT-003]", {
     # check with IOV model
     .s <- file.path(MODEL_DIR_X, "acop-iov") %>%
       read_model() %>%
@@ -147,7 +163,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
     expect_equal(res_str, ref_str)
   })
 
-  test_that("print.bbi_nonmem_summary .nrow argument", {
+  test_that("print.bbi_nonmem_summary .nrow argument [BBR-PRNT-003]", {
     # load a model summary
     .s <- file.path(MODEL_DIR_X, "acop-iov") %>%
       read_model() %>%
@@ -158,7 +174,7 @@ withr::with_options(list(bbr.bbi_exe_path = read_bbi_path()), {
   })
 
 
-  test_that("print.bbi_nonmem_summary .off_diag=TRUE", {
+  test_that("print.bbi_nonmem_summary .off_diag=TRUE [BBR-PRNT-003]", {
     .s <- file.path(MODEL_DIR_X, "example2_saemimp") %>%
       read_model() %>%
       model_summary()